SecondRankTensor => float instead of double
git-svn-id: https://advants.svn.sourceforge.net/svnroot/advants@263 fd2da0a4-d968-48df-9cb5-d8c496287cbf

+++ Examples/AverageTensorImages.cxx
@@ -10,7 +10,7 @@ int AverageTensorImages(unsigned int argc, char *argv[])
 {
 
   // typedef itk::Vector<float,6> TensorType;
-  typedef itk::SymmetricSecondRankTensor< double, 3 >  TensorType; 
+  typedef itk::SymmetricSecondRankTensor< float, 3 >  TensorType; 
 
   typedef itk::Image<TensorType, ImageDimension> ImageType;
   typedef itk::ImageRegionIteratorWithIndex<ImageType> IteratorType;
